requirements.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
asyncpg>=0.29.0

# Kafka
aiokafka>=0.10.0

# AI / Agents
openai>=1.76.0
openai-agents>=0.0.14

# Channel integrations
google-api-python-client>=2.100.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
twilio>=9.0.0

# Validation
pydantic[email]>=2.10.0

# Config / utils
python-dotenv>=1.0.0
